# docs/research/mvp/Agent_AI_Skill_Tree_Development/api/openapi.yaml
openapi: 3.1.0
info:
  title: HelixKnowledge Skill Graph API
  version: "1.0.0"
  summary: REST API for the self-growing Agent AI Skill Graph system.
  description: |
    Single source of truth for the skill-graph REST API, derived from
    `SPEC.md` §6 (API Specification), §4 (Data Models) and §7 (MCP tool
    parity).

    The Skill Graph stores versioned units of knowledge ("Skills") for AI
    CLI agents. Skills form a directed acyclic graph (DAG) of dependencies,
    are searchable via hybrid vector + keyword search, are auto-expanded
    to fill gaps, and are enriched with Evidence learned from real
    codebases.

    ## Authentication
    Every endpoint requires the `X-API-Key` header (see the `ApiKeyAuth`
    security scheme).

    ## Content negotiation
    This API negotiates **TOON** (Token-Oriented Object Notation,
    `application/toon` — see https://github.com/toon-format/toon) as the
    **primary** wire format for every request and response body, with
    **JSON** (`application/json`) always available as a **fallback**.
    Every operation that accepts or returns a body offers both media
    types against the same underlying schema; select via the
    `Content-Type` request header and the `Accept` response header
    (defaults to `application/json` when unspecified).
  contact:
    name: HelixKnowledge Skill Graph Team
    email: milos85vasic@gmail.com
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - ApiKeyAuth: []
...
paths:
  /skills:
    get:
      operationId: listSkills
      summary: List skills
      description: List skills with optional filters and pagination.
      tags: [Skills]
      parameters:
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/DomainFilter'
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/SearchFilter'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/AcceptFormatHeader'
      responses:
        '200':
          description: A page of skills.
          content:
            application/toon:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
              example:
                total: 2
                limit: 50
                offset: 0
                skills:
                  - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    name: "android.aosp.build-system"
                    version: "0.1.0"
                    title: "AOSP Build System (Soong, Make, Bazel)"
                    description: "Complete reference for Android build, Soong blueprints, Android.bp, etc."
                    content: "# AOSP Build System\n\n## Overview\n..."
                    metadata:
                      tags: ["android", "build", "soong", "bazel"]
                      domain: "android"
                      complexity: "intermediate"
                    status: "active"
                    created_at: "2026-01-10T09:00:00Z"
                    updated_at: "2026-06-01T14:22:31Z"
                  - id: "9c858901-8a57-4791-81fe-4c455b099bc9"
                    name: "python.basics"
                    version: "0.2.1"
                    title: "Python Fundamentals"
                    description: "Core Python language reference."
                    content: "# Python Fundamentals\n..."
                    metadata:
                      tags: ["python"]
                      domain: "python"
                      complexity: "beginner"
                    status: "validated"
                    created_at: "2026-01-05T09:00:00Z"
                    updated_at: "2026-01-05T09:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createSkill
      summary: Create skill
      description: |
        Create a new skill. Accepts a request body in TOON (`application/toon`,
        the default) or JSON (`application/json`). The canonical on-disk
        artifact remains the TOML+Markdown Skill Format (SPEC.md §4.2),
        retrievable from the git repository.
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/toon:
            schema:
              $ref: '#/components/schemas/SkillCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/SkillCreateRequest'
            example:
              name: "android.aosp.build-system"
              version: "0.1.0"
              title: "AOSP Build System (Soong, Make, Bazel)"
              description: "Complete reference for Android build, Soong blueprints, Android.bp, etc."
              content: "# AOSP Build System\n\n## Overview\nThe Android build system uses Soong (Android.bp), Make (Android.mk), and migrating to Bazel.\n"
              metadata:
                tags: ["android", "build", "soong", "bazel"]
                domain: "android"
                complexity: "intermediate"
              dependencies:
                requires: ["linux.kernel-modules", "python.basics", "make.basics"]
                extends: ["android.general"]
                recommends: ["bazel.advanced"]
              resources:
                - url: "https://source.android.com/docs/setup/build/building"
                  title: "Official Android Build Documentation"
                  resource_type: "official-doc"
                - url: "https://android.googlesource.com/platform/build/soong/+/refs/heads/main/README.md"
                  title: "Soong README"
                  resource_type: "code"
      responses:
        '201':
          description: Skill created.
          headers:
            Location:
              description: URL of the newly created skill.
              schema:
                type: string
                format: uri
          content:
            application/toon:
              schema:
                $ref: '#/components/schemas/Skill'
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: A skill with this name already exists.
          content:
            application/toon:
              schema:
                $ref: '#/components/schemas/Error'
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "conflict"
                message: "skill 'android.aosp.build-system' already exists"
                code: 409
        '422':
          description: Body parsed but failed validation (e.g. unresolved dependency name, invalid enum value).
          content:
            application/toon:
              schema:
                $ref: '#/components/schemas/Error'
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "unprocessable_entity"
                message: "dependency 'make.basics' does not exist"
                code: 422
...
  /skills/{name}/tree:
    get:
      operationId: getSkillTree
      summary: Get dependency tree
      description: Recursively resolve the dependency DAG rooted at the named skill.
      tags: [Skills]
      parameters:
        - $ref: '#/components/parameters/SkillNamePath'
        - name: depth
          in: query
          description: Maximum recursion depth (mirrors `autoexpand.max_depth`, default 5).
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
        - name: format
          in: query
          description: Response format. Equivalent to setting the `Accept` header.
          required: false
          schema:
            type: string
            enum: [json, toon]
            default: json
      responses:
        '200':
          description: Dependency tree.
          content:
            application/toon:
              schema:
                $ref: '#/components/schemas/SkillTreeNode'
            application/json:
              schema:
                $ref: '#/components/schemas/SkillTreeNode'
              example:
                skill:
                  id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  name: "android.aosp.build-system"
                  version: "0.1.0"
                  title: "AOSP Build System (Soong, Make, Bazel)"
                  status: "active"
                  created_at: "2026-01-10T09:00:00Z"
                  updated_at: "2026-06-01T14:22:31Z"
                relation_type: null
                depends_on:
                  - skill:
                      id: "9c858901-8a57-4791-81fe-4c455b099bc9"
                      name: "python.basics"
                      version: "0.2.1"
                      title: "Python Fundamentals"
                      status: "validated"
                      created_at: "2026-01-05T09:00:00Z"
                      updated_at: "2026-01-05T09:00:00Z"
                    relation_type: "requires"
                    depends_on: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /search:
    post:
      operationId: searchSkills
      summary: Hybrid vector + keyword search
      tags: [Search]
      requestBody:
        required: true
        content:
          application/toon:
            schema:
              $ref: '#/components/schemas/SearchRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            example:
              query: "android soong blueprint build rules"
              limit: 5
      responses:
        '200':
          description: Ranked search results.
          content:
            application/toon:
              schema:
                $ref: '#/components/schemas/SearchResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                query: "android soong blueprint build rules"
                took_ms: 42
                results:
                  - skill:
                      id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      name: "android.aosp.build-system"
                      version: "0.1.0"
                      title: "AOSP Build System (Soong, Make, Bazel)"
                      status: "active"
                      created_at: "2026-01-10T09:00:00Z"
                      updated_at: "2026-06-01T14:22:31Z"
                    score: 0.9123
                    match_type: "hybrid"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /registry/coverage:
    get:
      operationId: getCoverageReport
      summary: Coverage report
      tags: [Registry]
      parameters:
        - $ref: '#/components/parameters/DomainFilter'
      responses:
        '200':
          description: Coverage aggregated overall and per domain.
          content:
            application/toon:
              schema:
                $ref: '#/components/schemas/CoverageReport'
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageReport'
              example:
                generated_at: "2026-07-15T08:00:00Z"
                overall_coverage: 0.81
                coverage_threshold: 0.8
                by_domain:
                  - domain: "android"
                    coverage: 0.92
                    skill_count: 14
                  - domain: "python"
                    coverage: 0.55
                    skill_count: 6
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Static or rotating API key required by every `/api/v1` endpoint.
  parameters:
    SkillNamePath:
      name: name
      in: path
      required: true
      description: Dotted skill name, e.g. `android.aosp.build-system`.
      schema:
        type: string
      example: "android.aosp.build-system"
...
    StatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/SkillStatus'
    DomainFilter:
      name: domain
      in: query
      required: false
      description: Filter by `metadata.domain`.
      schema:
        type: string
    TagFilter:
      name: tag
      in: query
      required: false
      description: Filter by a single tag present in `metadata.tags`.
      schema:
        type: string
    SearchFilter:
      name: search
      in: query
      required: false
      description: Free-text filter over name/title/description (keyword only, not vector search).
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    OffsetParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    AcceptFormatHeader:
      name: Accept
      in: header
      required: false
      description: Set to `application/toon` (default) or `application/json` to choose the response representation.
      schema:
        type: string
        enum: [application/toon, application/json]
        default: application/json
...